Fixed Bug
Fixed a bug where the application would crash when you go to buy a vehicle. It would crash after trying to read from the binary file. This is because it was saving strings to the binary file and reading them as string which to my understanding the compiler didnt understand what it was reading. To fix this I converted the string types into char arrays.